NAS-132759: Refactor Installed Apps
