+++ src/app/pages/apps/components/installed-apps/installed-apps-list/installed-apps-list-bulk-actions/installed-apps-list-bulk-actions.component.html
@@ -1,0 +1,57 @@
+<div class="bulk-actions-container">
+  <div class="bulk-selected">
+    <span>{{ checkedApps().length }}</span>
+    <span>{{ 'Selected' | translate }}</span>
+  </div>
+
+  <div class="bulk-button-wrapper">
+    <label>{{ 'Bulk Actions' | translate }}</label>
+    <button
+      *ixRequiresRoles="requiredRoles"
+      mat-button
+      ixTest="bulk-actions-menu"
+      [matMenuTriggerFor]="menu"
+    >
+      {{ 'Select action' | translate }}
+      <ix-icon name="mdi-menu-down" class="menu-caret"></ix-icon>
+    </button>
+  </div>
+
+  <mat-menu #menu="matMenu">
+    <button
+      *ixRequiresRoles="requiredRoles"
+      mat-menu-item
+      ixTest="start-selected"
+      [disabled]="isBulkStartDisabled"
+      (click)="bulkStart.emit()"
+    >
+      <span>{{ 'Start All Selected' | translate }}</span>
+    </button>
+    <button
+      *ixRequiresRoles="requiredRoles"
+      mat-menu-item
+      ixTest="stop-selected"
+      [disabled]="isBulkStopDisabled"
+      (click)="bulkStop.emit()"
+    >
+      <span>{{ 'Stop All Selected' | translate }}</span>
+    </button>
+    <button
+      *ixRequiresRoles="requiredRoles"
+      mat-menu-item
+      ixTest="upgrade-selected"
+      [disabled]="isBulkUpgradeDisabled"
+      (click)="bulkUpgrade.emit()"
+    >
+      <span>{{ 'Upgrade All Selected' | translate }}</span>
+    </button>
+    <button
+      *ixRequiresRoles="requiredRoles"
+      mat-menu-item
+      ixTest="delete-selected"
+      (click)="bulkDelete.emit()"
+    >
+      <span>{{ 'Delete All Selected' | translate }}</span>
+    </button>
+  </mat-menu>
+</div>
