+++ src/app/pages/apps/components/installed-apps/installed-apps.component.html
@@ -33,182 +33,11 @@
   #masterDetailView="masterDetailViewContext"
   [selectedItem]="selectedApp"
 >
-  <ng-container master>
-    <div class="table-header">
-      <h2>{{ 'Applications' | translate }}</h2>
-
-      @if (hasCheckedApps) {
-        <div class="bulk-actions-container">
-          <div class="bulk-selected">
-            <span>{{ checkedAppsNames.length }}</span>
-            <span>{{ 'Selected' | translate }}</span>
-          </div>
-
-          <div class="bulk-button-wrapper">
-            <label>{{ 'Bulk Actions' | translate }}</label>
-            <button
-              *ixRequiresRoles="requiredRoles"
-              mat-button
-              ixTest="bulk-actions-menu"
-              [matMenuTriggerFor]="menu"
-            >
-              {{ 'Select action' | translate }}
-              <ix-icon name="mdi-menu-down" class="menu-caret"></ix-icon>
-            </button>
-          </div>
-
-          <mat-menu #menu="matMenu">
-            <button
-              *ixRequiresRoles="requiredRoles"
-              mat-menu-item
-              ixTest="start-selected"
-              [disabled]="isBulkStartDisabled"
-              (click)="onBulkStart()"
-            >
-              <span>{{ 'Start All Selected' | translate }}</span>
-            </button>
-            <button
-              *ixRequiresRoles="requiredRoles"
-              mat-menu-item
-              ixTest="stop-selected"
-              [disabled]="isBulkStopDisabled"
-              (click)="onBulkStop()"
-            >
-              <span>{{ 'Stop All Selected' | translate }}</span>
-            </button>
-            <button
-              *ixRequiresRoles="requiredRoles"
-              mat-menu-item
-              ixTest="upgrade-selected"
-              [disabled]="isBulkUpgradeDisabled"
-              (click)="onBulkUpgrade()"
-            >
-              <span>{{ 'Upgrade All Selected' | translate }}</span>
-            </button>
-            <button
-              *ixRequiresRoles="requiredRoles"
-              mat-menu-item
-              ixTest="delete-selected"
-              (click)="onBulkDelete()"
-            >
-              <span>{{ 'Delete All Selected' | translate }}</span>
-            </button>
-          </mat-menu>
-        </div>
-      }
-    </div>
-
-    <div class="item-search">
-      @if (!masterDetailView.showMobileDetails()) {
-        <ix-fake-progress-bar
-          class="loader-bar"
-          [loading]="isLoading()"
-        ></ix-fake-progress-bar>
-      }
-
-      <ix-search-input1
-        [maxLength]="100"
-        [disabled]="!dataSource.length"
-        [value]="filterString"
-        (search)="onSearch($event)"
-      ></ix-search-input1>
-    </div>
-
-    <div
-      class="sticky-header"
-      matSort
-      matSortActive="application"
-      matSortDirection="asc"
-      matSortDisableClear
-      (matSortChange)="sortChanged($event)"
-    >
-      <div class="app-header-row">
-        <div>
-          <span class="name-header">
-            @if (dataSource.length) {
-              <mat-checkbox
-                color="primary"
-                ixTest="select-all-app"
-                [checked]="allAppsChecked"
-                [indeterminate]="!allAppsChecked && !!selection.selected.length"
-                (change)="toggleAppsChecked($event.checked)"
-              ></mat-checkbox>
-            }
-          </span>
-        </div>
-        <div
-          [matColumnDef]="sortableField.Application"
-          [mat-sort-header]="sortableField.Application"
-        >
-          {{ 'Application' | translate }}
-        </div>
-        <div
-          [matColumnDef]="sortableField.State"
-          [mat-sort-header]="sortableField.State"
-        >
-          {{ 'Status' | translate }}
-        </div>
-        <div>{{ 'CPU' | translate }}</div>
-        <div>{{ 'RAM' | translate }}</div>
-        <div>{{ 'Block I/O' | translate }}</div>
-        <div>{{ 'Network' | translate }}</div>
-        <div
-          class="app-update-header"
-          [matColumnDef]="sortableField.Updates"
-          [mat-sort-header]="hasUpdates ? sortableField.Updates : null"
-          [disabled]="!hasUpdates"
-        >
-          <span>{{ 'Updates' | translate }}</span>
-          @if (hasUpdates) {
-            <ix-icon
-              class="has-updates-icon"
-              name="mdi-alert-circle"
-              matTooltipPosition="above"
-              [matTooltip]="'Updates available' | translate"
-            ></ix-icon>
-          }
-        </div>
-        <div>{{ 'Controls' | translate }}</div>
-      </div>
-    </div>
-
-    <div
-      matSort
-      matSortDisableClear
-      matSortActive="application"
-      matSortDirection="asc"
-      class="app-wrapper"
-      (matSortChange)="sortChanged($event)"
-    >
-      <div class="app-inner">
-        <div class="apps-rows">
-          @for (app of filteredApps; track app.name) {
-            <ix-app-row
-              tabindex="0"
-              [app]="app"
-              [stats]="getAppStats(app.name) | async"
-              [class.selected]="selectedApp.id === app.id"
-              [selected]="selection.isSelected(app.id)"
-              [job]="appJobs.get(app.name)"
-              (startApp)="start(app.name)"
-              (stopApp)="stop(app.name)"
-              (restartApp)="restart(app.name)"
-              (clickStatus)="openStatusDialog(app.name)"
-              (selectionChange)="selection.toggle(app.id)"
-              (click)="viewDetails(app)"
-              (keydown.enter)="viewDetails(app)"
-            ></ix-app-row>
-          }
-
-          @if ((dataSource.length && !filteredApps.length) || (!dataSource.length && !isLoading())) {
-            <div class="no-apps">
-              <ix-empty [conf]="entityEmptyConf"></ix-empty>
-            </div>
-          }
-        </div>
-      </div>
-    </div>
-  </ng-container>
+  <ix-installed-apps-list
+    master
+    [isMobileView]="masterDetailView.isMobileView()"
+    (toggleShowMobileDetails)="masterDetailView.toggleShowMobileDetails($event)"
+  ></ix-installed-apps-list>
 
   <ng-container detail-name>
     {{ selectedApp?.name }}
