+++ src/app/pages/apps/components/installed-apps/installed-apps.component.scss
@@ -1,155 +1,5 @@
-@import 'scss-imports/variables';
-
 :host {
   display: block;
-}
-
-.table-header {
-  align-items: flex-end;
-  color: var(--fg1);
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 12px;
-  min-height: 56px;
-
-  h2 {
-    align-items: center;
-    display: flex;
-    margin-top: 20px;
-    min-height: 36px;
-  }
-
-  .bulk {
-    align-items: center;
-    display: flex;
-    gap: 6px;
-  }
-}
-
-.app-wrapper {
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  overflow: auto;
-}
-
-.app-inner {
-  background-color: var(--bg2);
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  min-width: fit-content;
-
-  @media (max-width: $breakpoint-tablet) {
-    min-height: fit-content;
-    min-width: unset;
-    overflow-x: auto;
-  }
-}
-
-.no-apps {
-  color: var(--fg1);
-  display: flex;
-  justify-content: center;
-  margin: 15px;
-}
-
-ix-app-row,
-.app-header-row {
-  grid-template-columns: 5% minmax(18%, auto) 11% 4% 7% 11% 11% 13% 10%;
-
-  @media (max-width: $breakpoint-tablet) {
-    grid-template-columns: 45px auto 0 0 0 0 0 0;
-  }
-}
-
-.app-header-row {
-  align-items: center;
-  background: var(--bg1);
-  border-bottom: 1px solid var(--lines);
-  color: var(--fg2);
-  display: grid;
-  grid-gap: 8px;
-  min-height: 48px;
-  min-width: fit-content;
-
-  position: sticky;
-  top: 0;
-  width: 100%;
-  z-index: 1;
-
-  > div {
-    align-items: center;
-    display: flex;
-    font-weight: bold;
-    height: 100%;
-    justify-content: flex-start;
-    padding: 4px 0;
-
-    @media (max-width: $breakpoint-tablet) {
-      display: none !important;
-    }
-
-    &:first-child {
-      left: 0;
-      position: sticky;
-
-      @media (max-width: $breakpoint-tablet) {
-        display: block !important;
-      }
-    }
-
-    &:nth-child(2) {
-      @media (max-width: $breakpoint-tablet) {
-        display: flex !important;
-      }
-    }
-  }
-
-  .app-update-header {
-    align-items: center;
-    display: flex;
-    gap: 4px;
-
-    ix-icon {
-      color: var(--yellow);
-      font-size: 18px;
-      line-height: 1;
-      margin-left: 4px;
-    }
-  }
-}
-
-.bulk-selected {
-  align-items: center;
-  align-self: flex-end;
-  display: inline-flex;
-  font-size: 16px;
-  gap: 4px;
-  height: 36px;
-}
-
-.bulk-actions-container {
-  align-items: flex-end;
-  display: flex;
-  gap: 12px;
-}
-
-.bulk-button-wrapper {
-  display: flex;
-  flex-direction: column;
-
-  label {
-    color: var(--fg2);
-    font-size: 10px;
-    margin-bottom: 2px;
-  }
-
-  button {
-    background-color: var(--bg1);
-    border: 1px solid var(--lines);
-    font-size: 12px;
-  }
 }
 
 .global-update {
